+++ TombEngine/Specific/savegame/flatbuffers/ten_savegame_generated.h
@@ -6120,6 +6120,7 @@ struct SwarmObjectInfoT : public flatbuffers::NativeTable {
   bool on = false;
   std::unique_ptr<TEN::Save::Pose> pose{};
   int32_t room_number = 0;
+  int32_t target = 0;
   int32_t flags = 0;
 };
 
@@ -6131,7 +6132,8 @@ struct SwarmObjectInfo FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
     VT_ON = 4,
     VT_POSE = 6,
     VT_ROOM_NUMBER = 8,
-    VT_FLAGS = 10
+    VT_TARGET = 10,
+    VT_FLAGS = 12
   };
   bool on() const {
     return GetField<uint8_t>(VT_ON, 0) != 0;
@@ -6142,6 +6144,9 @@ struct SwarmObjectInfo FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
   int32_t room_number() const {
     return GetField<int32_t>(VT_ROOM_NUMBER, 0);
   }
+  int32_t target() const {
+    return GetField<int32_t>(VT_TARGET, 0);
+  }
   int32_t flags() const {
     return GetField<int32_t>(VT_FLAGS, 0);
   }
@@ -6150,6 +6155,7 @@ struct SwarmObjectInfo FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
            VerifyField<uint8_t>(verifier, VT_ON) &&
            VerifyField<TEN::Save::Pose>(verifier, VT_POSE) &&
            VerifyField<int32_t>(verifier, VT_ROOM_NUMBER) &&
+           VerifyField<int32_t>(verifier, VT_TARGET) &&
            VerifyField<int32_t>(verifier, VT_FLAGS) &&
            verifier.EndTable();
   }
@@ -6171,6 +6177,9 @@ struct SwarmObjectInfoBuilder {
   void add_room_number(int32_t room_number) {
     fbb_.AddElement<int32_t>(SwarmObjectInfo::VT_ROOM_NUMBER, room_number, 0);
   }
+  void add_target(int32_t target) {
+    fbb_.AddElement<int32_t>(SwarmObjectInfo::VT_TARGET, target, 0);
+  }
   void add_flags(int32_t flags) {
     fbb_.AddElement<int32_t>(SwarmObjectInfo::VT_FLAGS, flags, 0);
   }
@@ -6190,9 +6199,11 @@ inline flatbuffers::Offset<SwarmObjectInfo> CreateSwarmObjectInfo(
     bool on = false,
     const TEN::Save::Pose *pose = 0,
     int32_t room_number = 0,
+    int32_t target = 0,
     int32_t flags = 0) {
   SwarmObjectInfoBuilder builder_(_fbb);
   builder_.add_flags(flags);
+  builder_.add_target(target);
   builder_.add_room_number(room_number);
   builder_.add_pose(pose);
   builder_.add_on(on);
@@ -8602,6 +8613,7 @@ struct SaveGameT : public flatbuffers::NativeTable {
   std::vector<std::unique_ptr<TEN::Save::SwarmObjectInfoT>> spiders{};
   std::vector<std::unique_ptr<TEN::Save::SwarmObjectInfoT>> scarabs{};
   std::vector<std::unique_ptr<TEN::Save::SwarmObjectInfoT>> bats{};
+  std::vector<std::unique_ptr<TEN::Save::SwarmObjectInfoT>> locusts{};
   std::vector<std::unique_ptr<TEN::Save::DecalT>> decals{};
   std::vector<int32_t> flip_maps{};
   std::vector<int32_t> flip_stats{};
@@ -8671,42 +8683,43 @@ struct SaveGame FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
     VT_SPIDERS = 50,
     VT_SCARABS = 52,
     VT_BATS = 54,
-    VT_DECALS = 56,
-    VT_FLIP_MAPS = 58,
-    VT_FLIP_STATS = 60,
-    VT_FLIP_EFFECT = 62,
-    VT_FLIP_TIMER = 64,
-    VT_FLIP_STATUS = 66,
-    VT_CURRENT_FOV = 68,
-    VT_LAST_INV_ITEM = 70,
-    VT_ACTION_QUEUE = 72,
-    VT_SOUNDTRACKS = 74,
-    VT_CD_FLAGS = 76,
-    VT_VIDEO = 78,
-    VT_POSTPROCESS_MODE = 80,
-    VT_POSTPROCESS_STRENGTH = 82,
-    VT_POSTPROCESS_TINT = 84,
-    VT_ROPE = 86,
-    VT_PENDULUM = 88,
-    VT_ALTERNATE_PENDULUM = 90,
-    VT_VOLUMES = 92,
-    VT_GLOBAL_EVENT_SETS = 94,
-    VT_VOLUME_EVENT_SETS = 96,
-    VT_SCRIPT_VARS = 98,
-    VT_CALLBACKS_PRE_START = 100,
-    VT_CALLBACKS_POST_START = 102,
-    VT_CALLBACKS_PRE_END = 104,
-    VT_CALLBACKS_POST_END = 106,
-    VT_CALLBACKS_PRE_SAVE = 108,
-    VT_CALLBACKS_POST_SAVE = 110,
-    VT_CALLBACKS_PRE_LOAD = 112,
-    VT_CALLBACKS_POST_LOAD = 114,
-    VT_CALLBACKS_PRE_LOOP = 116,
-    VT_CALLBACKS_POST_LOOP = 118,
-    VT_CALLBACKS_PRE_USEITEM = 120,
-    VT_CALLBACKS_POST_USEITEM = 122,
-    VT_CALLBACKS_PRE_FREEZE = 124,
-    VT_CALLBACKS_POST_FREEZE = 126
+    VT_LOCUSTS = 56,
+    VT_DECALS = 58,
+    VT_FLIP_MAPS = 60,
+    VT_FLIP_STATS = 62,
+    VT_FLIP_EFFECT = 64,
+    VT_FLIP_TIMER = 66,
+    VT_FLIP_STATUS = 68,
+    VT_CURRENT_FOV = 70,
+    VT_LAST_INV_ITEM = 72,
+    VT_ACTION_QUEUE = 74,
+    VT_SOUNDTRACKS = 76,
+    VT_CD_FLAGS = 78,
+    VT_VIDEO = 80,
+    VT_POSTPROCESS_MODE = 82,
+    VT_POSTPROCESS_STRENGTH = 84,
+    VT_POSTPROCESS_TINT = 86,
+    VT_ROPE = 88,
+    VT_PENDULUM = 90,
+    VT_ALTERNATE_PENDULUM = 92,
+    VT_VOLUMES = 94,
+    VT_GLOBAL_EVENT_SETS = 96,
+    VT_VOLUME_EVENT_SETS = 98,
+    VT_SCRIPT_VARS = 100,
+    VT_CALLBACKS_PRE_START = 102,
+    VT_CALLBACKS_POST_START = 104,
+    VT_CALLBACKS_PRE_END = 106,
+    VT_CALLBACKS_POST_END = 108,
+    VT_CALLBACKS_PRE_SAVE = 110,
+    VT_CALLBACKS_POST_SAVE = 112,
+    VT_CALLBACKS_PRE_LOAD = 114,
+    VT_CALLBACKS_POST_LOAD = 116,
+    VT_CALLBACKS_PRE_LOOP = 118,
+    VT_CALLBACKS_POST_LOOP = 120,
+    VT_CALLBACKS_PRE_USEITEM = 122,
+    VT_CALLBACKS_POST_USEITEM = 124,
+    VT_CALLBACKS_PRE_FREEZE = 126,
+    VT_CALLBACKS_POST_FREEZE = 128
   };
   const TEN::Save::SaveGameHeader *header() const {
     return GetPointer<const TEN::Save::SaveGameHeader *>(VT_HEADER);
@@ -8785,6 +8798,9 @@ struct SaveGame FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
   }
   const flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *bats() const {
     return GetPointer<const flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *>(VT_BATS);
+  }
+  const flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *locusts() const {
+    return GetPointer<const flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *>(VT_LOCUSTS);
   }
   const flatbuffers::Vector<flatbuffers::Offset<TEN::Save::Decal>> *decals() const {
     return GetPointer<const flatbuffers::Vector<flatbuffers::Offset<TEN::Save::Decal>> *>(VT_DECALS);
@@ -8957,6 +8973,9 @@ struct SaveGame FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
            VerifyOffset(verifier, VT_BATS) &&
            verifier.VerifyVector(bats()) &&
            verifier.VerifyVectorOfTables(bats()) &&
+           VerifyOffset(verifier, VT_LOCUSTS) &&
+           verifier.VerifyVector(locusts()) &&
+           verifier.VerifyVectorOfTables(locusts()) &&
            VerifyOffset(verifier, VT_DECALS) &&
            verifier.VerifyVector(decals()) &&
            verifier.VerifyVectorOfTables(decals()) &&
@@ -9129,6 +9148,9 @@ struct SaveGameBuilder {
   void add_bats(flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>> bats) {
     fbb_.AddOffset(SaveGame::VT_BATS, bats);
   }
+  void add_locusts(flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>> locusts) {
+    fbb_.AddOffset(SaveGame::VT_LOCUSTS, locusts);
+  }
   void add_decals(flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::Decal>>> decals) {
     fbb_.AddOffset(SaveGame::VT_DECALS, decals);
   }
@@ -9276,6 +9298,7 @@ inline flatbuffers::Offset<SaveGame> CreateSaveGame(
     flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>> spiders = 0,
     flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>> scarabs = 0,
     flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>> bats = 0,
+    flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>> locusts = 0,
     flatbuffers::Offset<flatbuffers::Vector<flatbuffers::Offset<TEN::Save::Decal>>> decals = 0,
     flatbuffers::Offset<flatbuffers::Vector<int32_t>> flip_maps = 0,
     flatbuffers::Offset<flatbuffers::Vector<int32_t>> flip_stats = 0,
@@ -9348,6 +9371,7 @@ inline flatbuffers::Offset<SaveGame> CreateSaveGame(
   builder_.add_flip_stats(flip_stats);
   builder_.add_flip_maps(flip_maps);
   builder_.add_decals(decals);
+  builder_.add_locusts(locusts);
   builder_.add_bats(bats);
   builder_.add_scarabs(scarabs);
   builder_.add_spiders(spiders);
@@ -9411,6 +9435,7 @@ inline flatbuffers::Offset<SaveGame> CreateSaveGameDirect(
     const std::vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *spiders = nullptr,
     const std::vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *scarabs = nullptr,
     const std::vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *bats = nullptr,
+    const std::vector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> *locusts = nullptr,
     const std::vector<flatbuffers::Offset<TEN::Save::Decal>> *decals = nullptr,
     const std::vector<int32_t> *flip_maps = nullptr,
     const std::vector<int32_t> *flip_stats = nullptr,
@@ -9462,6 +9487,7 @@ inline flatbuffers::Offset<SaveGame> CreateSaveGameDirect(
   auto spiders__ = spiders ? _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>(*spiders) : 0;
   auto scarabs__ = scarabs ? _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>(*scarabs) : 0;
   auto bats__ = bats ? _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>(*bats) : 0;
+  auto locusts__ = locusts ? _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>>(*locusts) : 0;
   auto decals__ = decals ? _fbb.CreateVector<flatbuffers::Offset<TEN::Save::Decal>>(*decals) : 0;
   auto flip_maps__ = flip_maps ? _fbb.CreateVector<int32_t>(*flip_maps) : 0;
   auto flip_stats__ = flip_stats ? _fbb.CreateVector<int32_t>(*flip_stats) : 0;
@@ -9513,6 +9539,7 @@ inline flatbuffers::Offset<SaveGame> CreateSaveGameDirect(
       spiders__,
       scarabs__,
       bats__,
+      locusts__,
       decals__,
       flip_maps__,
       flip_stats__,
@@ -11270,6 +11297,7 @@ inline void SwarmObjectInfo::UnPackTo(SwarmObjectInfoT *_o, const flatbuffers::r
   { auto _e = on(); _o->on = _e; }
   { auto _e = pose(); if (_e) _o->pose = std::unique_ptr<TEN::Save::Pose>(new TEN::Save::Pose(*_e)); }
   { auto _e = room_number(); _o->room_number = _e; }
+  { auto _e = target(); _o->target = _e; }
   { auto _e = flags(); _o->flags = _e; }
 }
 
@@ -11284,12 +11312,14 @@ inline flatbuffers::Offset<SwarmObjectInfo> CreateSwarmObjectInfo(flatbuffers::F
   auto _on = _o->on;
   auto _pose = _o->pose ? _o->pose.get() : 0;
   auto _room_number = _o->room_number;
+  auto _target = _o->target;
   auto _flags = _o->flags;
   return TEN::Save::CreateSwarmObjectInfo(
       _fbb,
       _on,
       _pose,
       _room_number,
+      _target,
       _flags);
 }
 
@@ -12154,6 +12184,7 @@ inline void SaveGame::UnPackTo(SaveGameT *_o, const flatbuffers::resolver_functi
   { auto _e = spiders(); if (_e) { _o->spiders.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->spiders[_i] = std::unique_ptr<TEN::Save::SwarmObjectInfoT>(_e->Get(_i)->UnPack(_resolver)); } } }
   { auto _e = scarabs(); if (_e) { _o->scarabs.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->scarabs[_i] = std::unique_ptr<TEN::Save::SwarmObjectInfoT>(_e->Get(_i)->UnPack(_resolver)); } } }
   { auto _e = bats(); if (_e) { _o->bats.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->bats[_i] = std::unique_ptr<TEN::Save::SwarmObjectInfoT>(_e->Get(_i)->UnPack(_resolver)); } } }
+  { auto _e = locusts(); if (_e) { _o->locusts.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->locusts[_i] = std::unique_ptr<TEN::Save::SwarmObjectInfoT>(_e->Get(_i)->UnPack(_resolver)); } } }
   { auto _e = decals(); if (_e) { _o->decals.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->decals[_i] = std::unique_ptr<TEN::Save::DecalT>(_e->Get(_i)->UnPack(_resolver)); } } }
   { auto _e = flip_maps(); if (_e) { _o->flip_maps.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->flip_maps[_i] = _e->Get(_i); } } }
   { auto _e = flip_stats(); if (_e) { _o->flip_stats.resize(_e->size()); for (flatbuffers::uoffset_t _i = 0; _i < _e->size(); _i++) { _o->flip_stats[_i] = _e->Get(_i); } } }
@@ -12226,6 +12257,7 @@ inline flatbuffers::Offset<SaveGame> CreateSaveGame(flatbuffers::FlatBufferBuild
   auto _spiders = _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> (_o->spiders.size(), [](size_t i, _VectorArgs *__va) { return CreateSwarmObjectInfo(*__va->__fbb, __va->__o->spiders[i].get(), __va->__rehasher); }, &_va );
   auto _scarabs = _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> (_o->scarabs.size(), [](size_t i, _VectorArgs *__va) { return CreateSwarmObjectInfo(*__va->__fbb, __va->__o->scarabs[i].get(), __va->__rehasher); }, &_va );
   auto _bats = _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> (_o->bats.size(), [](size_t i, _VectorArgs *__va) { return CreateSwarmObjectInfo(*__va->__fbb, __va->__o->bats[i].get(), __va->__rehasher); }, &_va );
+  auto _locusts = _fbb.CreateVector<flatbuffers::Offset<TEN::Save::SwarmObjectInfo>> (_o->locusts.size(), [](size_t i, _VectorArgs *__va) { return CreateSwarmObjectInfo(*__va->__fbb, __va->__o->locusts[i].get(), __va->__rehasher); }, &_va );
   auto _decals = _fbb.CreateVector<flatbuffers::Offset<TEN::Save::Decal>> (_o->decals.size(), [](size_t i, _VectorArgs *__va) { return CreateDecal(*__va->__fbb, __va->__o->decals[i].get(), __va->__rehasher); }, &_va );
   auto _flip_maps = _fbb.CreateVector(_o->flip_maps);
   auto _flip_stats = _fbb.CreateVector(_o->flip_stats);
@@ -12290,6 +12322,7 @@ inline flatbuffers::Offset<SaveGame> CreateSaveGame(flatbuffers::FlatBufferBuild
       _spiders,
       _scarabs,
       _bats,
+      _locusts,
       _decals,
       _flip_maps,
       _flip_stats,